Consolidate CLI
