+++ tests/test_main.py
@@ -8,15 +8,14 @@ import csv
 from click.testing import CliRunner
 
 import newlinejson as nlj
-from newlinejson import tool
+from newlinejson.__main__ import main
 
 
 def test_csv2nlj(tmpdir, dicts_csv_path, dicts_path):
     outfile = str(tmpdir.mkdir('test').join('out.json'))
 
-    result = CliRunner().invoke(tool.csv2nlj, [
-        dicts_csv_path,
-        outfile
+    result = CliRunner().invoke(main, [
+        'csv2nlj', dicts_csv_path, outfile
     ])
     assert result.exit_code == 0
     with nlj.open(dicts_path) as expected,\
@@ -28,9 +27,8 @@ def test_csv2nlj(tmpdir, dicts_csv_path, dicts_path):
 def test_nlj2csv(tmpdir, dicts_path):
     outfile = str(tmpdir.mkdir('test').join('out.json'))
 
-    result = CliRunner().invoke(tool.nlj2csv, [
-        dicts_path,
-        outfile
+    result = CliRunner().invoke(main, [
+        'nlj2csv', dicts_path, outfile
     ])
     assert result.exit_code == 0
     with nlj.open(dicts_path) as expected, open(outfile) as actual:
